requirements.txt
Flask==1.1.1
flask-cors==3.0.8
python-dotenv==0.12.0
pandas==1.0.1
scikit-learn==0.22.2
gunicorn==20.0.4
